backend/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
pymysql 
passlib[bcrypt] 
python-jose 
authlib 
python-dotenv
pydantic[email]
httpx


torch 
numpy 
pypdf 
python-docx 
openai-whisper 
pyttsx3 
SpeechRecognition 
gTTS  
sentence-transformers 
scikit-learn 
opencv-python 
pillow 
torchvision
